Move SequenceMap and SequenceSet to
