+++ platform-sdk/consensus-model/build.gradle.kts
@@ -2,6 +2,16 @@
 plugins {
     id("org.hiero.gradle.module.library")
     id("org.hiero.gradle.feature.publish-artifactregistry")
+    id("org.hiero.gradle.feature.test-timing-sensitive")
 }
 
 description = "Consensus Model"
+
+timingSensitiveModuleInfo {
+    requires("com.swirlds.common")
+    requires("com.swirlds.common.test.fixtures")
+    requires("org.hiero.base.concurrent")
+    requires("org.hiero.base.utility.test.fixtures")
+    requires("org.junit.jupiter.api")
+    requires("org.junit.jupiter.params")
+}
