+++ platform-sdk/consensus-model/src/main/java/module-info.java
@@ -9,6 +9,9 @@ module org.hiero.consensus.model {
     exports org.hiero.consensus.model.io.streams;
     exports org.hiero.consensus.model.node;
     exports org.hiero.consensus.model.notification;
+    exports org.hiero.consensus.model.sequence;
+    exports org.hiero.consensus.model.sequence.map;
+    exports org.hiero.consensus.model.sequence.set;
     exports org.hiero.consensus.model.state;
     exports org.hiero.consensus.model.status;
     exports org.hiero.consensus.model.stream;
@@ -18,6 +21,7 @@ module org.hiero.consensus.model {
     requires transitive com.hedera.node.hapi;
     requires transitive com.hedera.pbj.runtime;
     requires transitive org.hiero.base.concurrent;
+    requires transitive org.hiero.base.utility;
     requires com.swirlds.base;
     requires static transitive com.github.spotbugs.annotations;
 }
