+++ platform-sdk/consensus-model/src/main/java/org/hiero/consensus/model/sequence/set/AbstractSequenceSet.java
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.sequence.set.internal;
+package org.hiero.consensus.model.sequence.set;
 
-import com.swirlds.platform.sequence.map.SequenceMap;
-import com.swirlds.platform.sequence.set.SequenceSet;
+import org.hiero.consensus.model.sequence.map.SequenceMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.List;
 import java.util.Objects;
