+++ platform-sdk/consensus-model/src/main/java/org/hiero/consensus/model/sequence/set/SequenceSet.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.sequence.set;
+package org.hiero.consensus.model.sequence.set;
 
-import com.swirlds.platform.sequence.Shiftable;
+import org.hiero.consensus.model.sequence.Shiftable;
 import java.util.List;
 import java.util.function.Consumer;
 import org.hiero.base.utility.Clearable;
