+++ platform-sdk/swirlds-platform-core/build.gradle.kts
@@ -52,8 +52,6 @@ timingSensitiveModuleInfo {
     requires("com.swirlds.platform.core")
     requires("com.swirlds.platform.core.test.fixtures")
     requires("org.hiero.base.concurrent")
-    requires("org.hiero.base.utility.test.fixtures")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
-    requires("org.junit.jupiter.params")
 }
